+++ recipes/mod_auth_anonymous.rb
@@ -18,7 +18,11 @@
 # limitations under the License.
 #
 
-include_recipe 'iis'
+iis_install 'install IIS' do
+  source node['iis']['source']
+  install_method node['iis']['windows_feature_install_method']
+  start_iis true
+end
 
 iis_section 'unlocks anonymous authentication control in web.config' do
   section 'system.webServer/security/authentication/anonymousAuthentication'